make sure all posture checks on windows are prepared by the daemon
